Bug fix: if ﻿﻿﻿﻿antworld/server/Cell.java.getFoodUnits() is called where there are no object in the cell, then the function returns 0 (it used to through a null pointer exception). Thanks  Sherman  for pointing this out.
All nest cells now have zero elevation change.

Healing can now be done using water OR food (it used to only use food). In the next, the default is water, but if there is no water, then food is automatically used

Food spawn locations have been set to use the placement algrothom that will be used in the actual grading